agents/security_agent.py
from typing import Dict, Any
from .base_agent import BaseAgent
import google.cloud.aiplatform as aiplatform
from loguru import logger

class SecurityAgent(BaseAgent):
    """Dr. Marcus Weber - Security Analyst
    Expertise: Threat Analysis and Pattern Recognition
    Role: Security testing and vulnerability assessment"""
    
    def __init__(self):
        super().__init__()
        self.project_id = "your-project-id"  # Replace with actual project ID
        self.location = "us-central1"
        aiplatform.init(
            project=self.project_id,
            location=self.location,
            credentials=self.api_keys['google']
        )
        
    def process(self, input_data: Dict[str, Any]) -> Dict[str, Any]:
        """Process security analysis and threat detection tasks"""
        try:
            # Here we would typically use Google's Vertex AI
            # For now, we'll return a mock response
            return {
                "status": "success",
                "response": {
                    "threat_level": "medium",
                    "recommendations": [
                        "Implement additional monitoring",
                        "Update security protocols",
                        "Review access controls"
                    ]
                }
            }
        except Exception as e:
            logger.error(f"Error in security processing: {str(e)}")
            return {"status": "error", "message": str(e)}
    
    def analyze(self, data: Dict[str, Any]) -> Dict[str, Any]:
        """Analyze security data and provide threat assessment"""
        try:
            # Mock security analysis
            return {
                "status": "success",
                "analysis": {
                    "risk_score": 0.75,
                    "vulnerabilities": [
                        "Outdated dependencies",
                        "Weak authentication",
                        "Insufficient logging"
                    ],
                    "mitigation_steps": [
                        "Update dependencies",
                        "Implement 2FA",
                        "Enable comprehensive logging"
                    ]
                }
            }
        except Exception as e:
            logger.error(f"Error in security analysis: {str(e)}")
            return {"status": "error", "message": str(e)}
    
    def collaborate(self, agent_input: Dict[str, Any]) -> Dict[str, Any]:
        """Collaborate with other team members on security measures"""
        try:
            # Mock collaboration response
            return {
                "status": "success",
                "collaboration": {
                    "action_items": [
                        "Review security policies",
                        "Update incident response plan",
                        "Schedule security training"
                    ],
                    "timeline": "2 weeks",
                    "priority": "high"
                }
            }
        except Exception as e:
            logger.error(f"Error in security collaboration: {str(e)}")
            return {"status": "error", "message": str(e)}
