requirements.txt
openai>=1.0.0
anthropic>=0.7.0
groq>=0.3.0
google-cloud-aiplatform>=1.38.0
cohere>=4.32
python-dotenv>=1.0.0
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0
python-multipart>=0.0.6
loguru>=0.7.2
numpy>=1.24.0
pandas>=2.1.0
scikit-learn>=1.3.0
requests>=2.31.0
